refac: discount model 분리, uscart에 연결
